pyproject.toml
[project]
name = "e2e-testing-agent"
version = "2.8.0"
description = "Autonomous E2E Full-Stack Testing Agent powered by Claude"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Argus Team", email = "team@heyargus.com" }
]

dependencies = [
    # Core AI
    "anthropic>=0.75.0,<1.0.0",  # Latest v0.75.x
    "google-genai>=1.0.0",  # Gemini Computer Use SDK
    "google-generativeai>=0.8.0",  # Gemini standard API
    "langgraph>=1.0.5,<2.0.0",  # Latest stable v1.x
    "langchain-anthropic>=1.3.0,<2.0.0",  # Latest v1.x
    "langchain-core>=1.2.5,<2.0.0",  # Latest v1.x

    # Web automation (multi-framework support)
    "playwright>=1.48.0",
    "selenium>=4.25.0",
    "webdriver-manager>=4.0.0",
    "httpx>=0.27.0",
    "websockets>=13.0",

    # Data validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.1.0",  # Required for EmailStr

    # Database
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",

    # LangGraph Checkpointing (durable execution)
    "langgraph-checkpoint>=2.0.0",
    "langgraph-checkpoint-postgres>=2.0.0",
    "psycopg[binary]>=3.1.0",

    # MCP
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",

    # Docker
    "docker>=7.0.0",

    # Utilities
    "rich>=13.9.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.8.0",
    "pillow>=10.4.0",
    "numpy>=1.26.0",
    "imagehash>=4.3.0",  # Perceptual image hashing
    "scikit-image>=0.22.0",  # SSIM and image metrics

    # API server (for webhooks)
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.6",
    "sse-starlette>=2.0.0",

    # Supabase integration
    "supabase>=2.0.0",
    "aiohttp>=3.9.0",

    # Email
    "aiosmtplib>=3.0.0",

    # Embeddings for memory store
    "openai>=1.0.0",

    # Security / JWT
    "PyJWT[crypto]>=2.8.0",  # JWT with RS256 support for Clerk
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "ruff>=0.7.0",
    "mypy>=1.12.0",
]

[project.scripts]
e2e-agent = "e2e_testing_agent.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]  # Removed N (naming) - too strict for existing codebase
ignore = [
    "E402",  # Module level import not at top - needed for conditional imports
    "E501",  # Line too long - relax for readability
    "E721",  # Do not compare types - sometimes needed
    "E722",  # Do not use bare 'except' - acceptable in test mocks
    "E741",  # Ambiguous variable name (l, O, I) - needed for LAB color space
    "F401",  # Imported but unused - many are for availability checking (HAS_X flags)
    "F841",  # Local variable assigned but never used - often intentional in tests
    "UP031", # Use format specifiers - existing code uses % formatting
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E501"]  # Allow long lines in tests

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/migrations/*",
    "*/.venv/*",
    "*/venv/*",
]
parallel = true

[tool.coverage.paths]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
fail_under = 95
show_missing = true
precision = 2

[tool.coverage.html]
directory = "htmlcov"

[tool.coverage.xml]
output = "coverage.xml"

[tool.coverage.json]
output = "coverage.json"

[tool.mypy]
python_version = "3.12"
strict = true
